backend/travel/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
PyJWT==2.8.0
stripe==9.4.0
psycopg2==2.9.7
gunicorn==20.1.0